Initial scripts to configure "baremetal" VMs
To test deployment with Ironic we'll need some VMs to act as dummy
baremetal nodes, similar to how we test tripleo.

Since the use-case is similar I've copied some script from the
https://github.com/imain/ocp-doit/ repo which is used for dev
testing of OCP deployed on OpenStack.

I've also reused some tripleo-quickstart roles to do the virt setup,
but with a configuration to enable privileged virt (this way VMs
survive host reboot and are visible by default via virt-manager and
virsh etc).

Steps to test (on a CentOS7 box) are:

- Ensure you have passwordless sudo
- 01_install_requirements.sh
- 02_configure_host.sh

All other scripts are WIP and/or initial imports, and also note
02_configure_host.sh doesn't yet include the role for the vbmc
setup (that will be added pending further local testing).