First pass creating bootstrap VM & ignition config
This generates the ignition configs with the installer
built to enable the (currently dev-only) libvirt backend,
then we generate ignition configs and launch the bootstrap
VM with the RHCOS image.

This is a first pass that creates ignition configs then launches
a bootstrap VM, with xml files based on those generated when using
the openshift-install libvirt backend.

Several TODO items exist:
- FIXME_ADDSECRET_HERE in ocp_install_env.sh, need to either implement
  per-user configs or use a default here
- Figure out why the libvirt backend always launches a master, if we
  can disable this we can probably remove the manual libvirt wrangling
- The manual libvirt stuff start the VM but something is wrong with
  the volume attachment for the bootstrap.ign (need to attach-disk?)