Add 06_deploy_masters.sh
This is a first-pass script which shows how to deploy the three
master nodes via Ironic.  Later this is likely to be automated
on the bootstrap VM (ref https://github.com/metalkube/dev-scripts/pull/37)
but for now having a script makes it easy to hack on.

Also rename 05_run_ocp.sh to 05_deploy_bootstrap_vm.sh
since that isn only doing the first step of the OCP deploy.