Ignition patching system
This patch introduces a systematic way of patching ignition files and
provides a utility method for its usage.

All your patches to ignition files should follow jsonpatch rfc 6902.
You should place your patch json files in:

    ignition_patches/kind_of_machine

Then, in the code that starts the machine, you should run:

    apply_ignition_patches bootstrap "$IGN_FILE"

where bootstrap is the kind of machine from the possible:

- bootstrap
- masters
- workers

and IGN file is the ignition file that will be served to the machine.
To run this, you should source utils.sh