bootstrap node keepalived config for master VIP
With this patch, the bootstrap node will start keepalived and bring up
the VIP that is resolved by the DNS server.

In order to prevent timing out on NetworkManager-wait-online, I moved
ironic to run as a systemd service that gets its interface configured by
bootstrap.ign rather than waiting for the baremetal network ssh to work
(we still wait for that to copy the coreos image and related config
without adverse effects to systemd unit starting).