Download ironic-inspector container and allow introspection to run
This gets the the ironic-inspector container and runs it. It also
changes the boot.ipxe file used for deployment to include
inspector.ipxe for introspection. If a deployment is not
being done, i.e. Ironic has not created a file in
/var/www/html/pxelinux.cfg/<MAC>, then the inspector.ipxe
file will be chainloaded to cause the IPA to run introspection
and return data to Ironic Inspector.