Allow disabling master/worker VMs for baremetal testing
When testing with real baremetal master/worker nodes we only
need the bootstrap VM, so we can disable some of the quickstart
roles by setting "NODES_PLATFORM" to something other than "libvirt"
which is the default.

NODES_FILE can also be overridden to point to your manually created
alternative to $WORKING_DIR/ironic_nodes.json (or you can just
copy it to that location).