Also remove ocp/install-config.yaml during cleanup
In situations where PULL_SECRET is either bad or broken, the
contents of this will still be written to install-config.yaml.
It's still assumed to be valid.  Since we source config_*.sh
anyway before running any of the top-level cleanups, we can
expect to just write this value to ocp/install-config anyway.