Reduce the RHCOS disk size
The RHCOS image disk size is 16G so when IPA is writing
the qcow2 to disk during deployment, it has to write 16G (at least)
to the master disk. This happens 3 times, so on a virtual env we
are writting 50G during each deploy. The partitions on the RHCOS
image are only about 3.3G so we can safely reduce the disk size and
therfore reduce the amount of data ironic writes. When the master nodes
boot growfs runs and expands the filesyatem to the size available
on the master node, so we end up using all the space available anyways.

This patch in local tests takes 3 minutes off the deploy time for
the master nodes, so should speed up development time, in the
longer term we should look into changing the image upstream.