Update install-config (#86)
Update to v1beta3 install-config.yaml

Use the defaults for everything except the pieces of
information that are required or that we want to override.

We were setting the machine pool replicas to null to try
and get "create cluster" to launch a bootstrap VM but no
masters. According to the "FIXME: this doesn't work", it
looks like we're not depending on this right now.

The connection URI isn't actually used for generating ignition
configs.