Create link to the latest RHCOS image
Create a link(redhat-coreos-maipo-latest.qcow2) pointing to the
the RHCOS image that we should be using. Also precalcualte the
md5sum of this image and expose it over http. This will allow
a client to reference the image to deploy and get its md5sum,
without having to download the data to calculate it.