Cleanup ssh known hosts in ocp_cleanup.sh
This is helpful to avoid warnings when ssh is used manually
from the host to access the deployed nodes.

We might also want to clean this in 07_deploy_masters.sh when
we detect that the VIP has failed over from the bootstrap node,
but I've not considered that in this patch.