Prepare repository for CI
Add a script that can be triggered on github pull request,
For now just call make but will be expanded in the future.
Also prevent the contents of PULL_SECRET from being output to
the console and add a catchall in run_ci.sh incase it creeps
back in future.