Create default libvirt storage pool
It turns out that the default libvirt storage pool is created by tools
like virt-install or virt-manager:

https://www.redhat.com/archives/libvir-list/2008-August/msg00179.html

If we're no longer using those tools, we'll have to do it
ourselves. Since openshift-install already requires the pool to exist,
it seems reasonable (at least for now) for kni-install to also require
it. See:

https://github.com/openshift/installer/blob/master/docs/dev/libvirt-howto.md#configure-default-libvirt-storage-pool