Use a higher number pci bus for the provisioning nic
This change sets a higher pci bus number for the provisioning nic
device(default is 0x00) in order for the OS to detect this nic as
the 2nd network interface. This is so we match the networking
prerequisites[1] that we document for baremetal environments.

Note: we need to keep the provisioning nic first in the libvirt
domain definition to allow pxe booting via virtualbmc. Currently
virtualbmc only supports setting the domain's `network` boot device
which always sets the 1st nic as the boot device.

[1] https://docs.google.com/document/d/1qTBuV5uRTDd1rh7TByFFTbFfVPSEKD44sN6IKHojgYA/edit