Use ssh with ConnectTimeout=5
I've seen the ssh to check if the master API is ready
(in 07_deploy_masters.sh) hang and have to be killed for the
script to move one, presumably this was something to do
with the VIP moving from the bootstrap node to the master.
Add a timeout to all ssh commands to prevent this from happening,
also create a common SSH env variable for all ssh commands.