Add extra manifests during 06_deploy_bootstrap_vm.sh
We add a create manifests step before generating the ignition
configs, because the assets in kni-installer will read all files
in the location, so we can easily patch in the extra manifests
needed.

In future it may be cleaner to add the manifest assets directly
to kni-installer, but for now this should be a simple way to
iterate on testing the manifests.

Note that the index prefix was changed to align with the existing
manifests generated by kni_installer, I'm assuming using a larger
index here will result in the desired serialization so these are
applied post-openshift things, testing needed to confirm this.