Fix problems with ci script
The pull secret contains "auths" search for the extra
"s" as without it this catchall if filtering random lines
in the console that don't need to be filtered.

In particular it has been filtering out the calls to
oc --config ocp/auth/kub....

Move the definition of CONFIG
This needs to happen before we import common.sh as we
check for its existence in there, CI only works at the
moment becase we also have it defined in the jenkins job
definition.

Find files to remove with find, the wildcard in the path
wasn't working, as /opt/libvirt-images was owned by root.

Also set pipefile to if make fails it will cause the script
to fail.