Move provisioing network setup to 02_configure_host.sh
We have code for the BM case that uses it in 02_configure_host.sh
so we need to create it before we try and add iptables rules to it.

Change the setup of the provisioning network and how PRO_IF
is added to it so it persists.

Also allow dhcp on the provisioing network.