Do not rely on virsh net-dhcp-leases to determine the bootstrap VM IP
For baremetal use case the DHCP server on the baremetal network is
external to the provisioning host so virsh net-dhcp-leases will not
work to determine the bootstrap VM IP address. Instead we can use the
API VIP(known prerequisite) to SSH into the bootstrap VM and obtain the
local interface address.