Stop using /etc/resolv.conf to get CLUSTER_DOMAIN
Using the "search" line in /etc/resolv.conf requires requires
that nodes be renamed to match the cluster they are part of. This
mean we can't split up a set of node into two envirments without
changing the DNS entry for each node.

Add and use a tool that gets the CLUSTER_DOMAIN from
the api url in /etc/kubernetes/kubeconfig