Make master node mutation be script safe
oc get is for humans and is not script safe. Use a "-o" so we get something
script safe.

Drop all usage of --config since it was used inconsitenctly and the kubeconfig
was already set by common.sh

Stop doing bash loops for operations which are selectable across labels
by kubernetes.