add a script to register hosts
Add an optional step 11 to register hosts for the masters and workers.

Masters are registered with a machine reference, to prevent them from
showing up as available to be provisioned.

Workers are registered without a machine reference, so they are
available to be provisioned.

Add the missing machine and boot-mac arguments to make-bm-workers.