Reference correct domain env var
CLUSTER_DOMAIN was removed in 8474c871f3904c11d76ca1ac65997a5c4a3c29a5
but this reference wasn't updated. As a result, we've been querying
for "ns1." when starting coredns. This seems to work fine in my
virt environment, but I just debugged a system where that was
resulting in an empty DNS_VIP. This caused coredns to crash if it
tried to forward a query because there was nothing in resolv.conf.
The grep -v "" removed all lines, not just the one we wanted.