Fix rhcos image caching in CI
The problem is we set `FILESTOCACHE` before we actually know the image
name - we get that from the installer and at the beginning of CI we
haven't fetched the installer yet. `$RHCOS_IMAGE_FILENAME_OPENSTACK`
isn't set... :chicken: :egg:

This fetches the last known image from the cache, and saves back the
current image once it's known at the end.