Don't get the RHCOS image URL from kni-installer
There's no straightforward way to get the installer pinned RHCOS
version from an installer binary or a release payload, so let's not
become reliant on this. We can easily pin dev-scripts to the correct
RHCOS image each time we pin to a new release payload.

Also, ultimately we need to get away from this way of caching a
pre-defined RHCOS image version - we instead need a way of caching
whatever version is driven by either openshift-install or machine
config upgrades.