Only wait for worker-0 in the default virt config.
We automate waiting for a worker to come up and adding IPs to it for the
default virt configuration.  This is a helpful step for the common dev setup,
and it also runs in CI. For any other env, we just skip this, because we
can't automatically figure out the mapping between Machines and Nodes in
other cases, and must rely on running the link-machine-and-node.sh manually.

Related issues: #421 and #417