Add 3 second delay before linking machines and nodes
The preceeding steps leading up to running link-machine-and-node.sh
via add-machine-ips.sh might create a condition on slower systems
or systems with resource contention to not start the machine API
soon enough for the link-machine-and-node.sh steps to succesfully
interact with an API that hasn't yet come up.  The result of this
failure is "Failed connect to localhost:8001; Connection refused".