10_deploy_rook.sh: set mgr_debug_aggressive_pg_num_changes during setup
This is a workaround until https://github.com/ceph/ceph/pull/27876
merges and makes it into the ceph build. It makes the mgr make aggressive
pgp_num changes so that the ceph health settles quickly after adjusting
the pg_num for new pools.

We can remove this later once the above ceph fixes are in place.

Signed-off-by: Sage Weil <sage@redhat.com>