Move roles/environment to libvirt role
This only configured the libvirt networking, and this should really
happen after the install of libvirt anyway, so lets move the tasks
to the libvirt role, and decouple the install, network setup and
vm setup tasks.

Note that the use of the platform variable has been renamed to
vm_platform to avoid breaking dev-scripts users that rely on the
existing behavior, but this is now optional for environments where
all libvirt things should be configured.

Fixes: #514