Remove machine-api-controllers restart.
This was a workaround that is no longer necessary since we're now
using a new enough release of OpenShift that includes the fix.

The upstream fix was here:
https://github.com/metal3-io/cluster-api-provider-baremetal/pull/58

and it was backported here:
https://github.com/openshift/cluster-api-provider-baremetal/pull/21

The current release we're using can be checked using:
oc adm release info quay.io/openshift-release-dev/ocp-release:4.1.0-rc.3 -a ~/pull-secret.json --commits | grep baremetal
  baremetal-machine-controllers https://github.com/openshift/cluster-api-provider-baremetal 8a27778353fb5f236593cdee112d4542e6f96929

This commit is past when the fix was merged into
openshift/cluster-api-provider-baremetal.