Add kni-installer to projects that get moved to $GOPATH
In #558, I added the ability to run the installer from git, assuming
that CI would check out the repo to $GOPATH, but that was an incorrect
assumption, it gets checked out to $HOME. We need it in the list of
projects that gets moved to $GOPATH, if the path exists.

This does not affect the behavior of installation from release payloads
by default.