Also delete the interface
When cleaning up the interface, remove the interface itself from the system.
The interface becomes unconfigured, though it won't delete the interface
from the system.  Ordinarily the next step would be to reboot the system,
at which time the interface would be deleted from the system.  However, not
rebooting the system could result in errors similar to: "error creating
bridge interface baremetal: File exists"