Updates to sync_repo_and_patch to support reproducible builds
For the CI workflow it is important to be able to reproduce builds.
This patch will do two things.  if a repos variable is defined with
repo/path/on/disk;SHA entries it will check out that specific
version instead master.  Second, if master is checked out this
will record the current HEAD commit.

Our CI workflow will cat the recorded_repos entry to config_$user.sh
which we then archive as an artifact.

This should allow us to re-run builds and get the same git repos.