Drop rook and kubevirt scripts.
These changes are to trim dev-scripts a bit further back to its
original intended scope, which is a development platform for bare
metal IPI support for the openshift installer.

These scripts made sense for demo purposes, but now that we're past
that, we should focus on evolving to the desired deployment flow.
We assume that the deployment of CNV (kubevirt++) or OCS (rook++) are
done as day 2 operations, presumably via an operator managed by OLM.

Kubevirt or rook developers may still want to manually deploy their
components on a dev-scripts cluster for development purposes,
bypassing OLM, or as a way to test things before the OLM path is
ready, but let's keep that separate from these scripts.  If desired,
those could be placed in a new repo that tells someone to deploy a
dev-scripts cluster as the first step.