cleanup leaked bootstrap VM/volume after failure
Reviving the PR initially pushed via #309 because I now see that
we leak bootstrap VMs/volumes when kni-installer fails while running
terraform.

This appears to be a limitation and/or design flaw with terraform as
it's not persisting state of created resources when exiting due to errors.

This is inconvenient when developing changes to the installer and/or
terraform pieces, so lets do some automated cleanup here.

Co-Authored-By: Alexander Chuzhoy achuzhoy@redhat.com