Extend Ironic conductor heartbeat timeout
In single node virtual environments, the system can get so overloaded
during deployment of the masters that the heartbeat times out, which
causes #617. This change overrides the heartbeat timeout to change
it from the default of 60 seconds to 120 seconds. In my experience,
this is sufficient to prevent the timeouts.

Note that this is making use of the environment driver[0] in oslo.config
for setting the value. I don't think we want to change the value
in the container config since it's primarily for virtual dev
environments.

0: https://docs.openstack.org/oslo.config/latest/reference/drivers.html#environment