Add sanity check to vip-prepender
We recently had a deployment where the DNS VIP was not properly added
to resolv.conf. There isn't enough detail in the logs from the script
to determine what went wrong, so this change adds some more logging
and a sanity check that we actually did the thing we were trying to do.