Demonstrate installer early exit is not necessary.
The only remaining change we have running after exiting early from the
install process is adding IPs to the master Machines.  This is to
enable auto approval of CSRs.  We also have a cron job that does this,
so it's not necessary to do this in the middle of the install.  The
change moves it to post-install.

Related issues:
  https://github.com/openshift-metal3/kni-installer/issues/60
  https://github.com/openshift-metal3/dev-scripts/issues/260
  https://github.com/metal3-io/baremetal-operator/issues/242