Attempt at host-recorder/smart exporter pod
In order to allow us to both identify deployed hosts
and match them up with machiens that we manage via metal3,
and obtain smart data for the hard disks, we need to add
some additoinal services.

The first service exposes the smart data.

The second service will post data to the ironic-inspector-image
container running as part of the baremetal operator.