Add dnsVIP to install-config
In order to generate the configuration for keepalived and coredns
(see keepalived.sh and coredns.sh) we currently determine the
DNS VIP that has been configured by querying the
ns1.$cluster_name.$base_domain record. This implies that users
must configure this record before beginning installation. In order to
reduce the pre-installation configuration requirements, we can
instead require the cluster administrator to provide the DNS VIP
as part of the configuration supplied in install-config.

Before proposing a change to the installer to support the DNS
VIP configuration, add it here to the install-config we generate,
so that the installer change can be tested.