Remove some cert approval related hacks.
This patch cleans up some hacks that are no longer needed.  CSRs for
masters are automatically approved by a service runningon the
bootstrap node during the bootstrap process.

Workers is where we are still missing some pieces.  Move the fix-certs
cron job to the register_hosts script, which is where worker
deployment is kicked off.

Also remove the add-machine-ips and link-machine-and-node scripts.
cluster-api-provider-baremetal should now be automatically populating
the addresses field of Machines based on introspection data it gets
out of the BareMetalHost object for workers.  Removing these scripts
ensures we only rely on that code, and not a dev-script hack.

The reigster_hosts script no longer waits for the worker to come up,
since it's not required and the cluster should already be functional
by this point.  The run_ci script was updated to ensure the worker
comes up to still make sure this case is tested.