Add validation for number of nodes in NODES_FILE
Some folks have seen a confusing error like:

Terraform variables: Unknown BMC type 'null' for address null://null\""

It turns out this is because the NODES_FILE created outside
of dev-scripts contains only two nodes, so the jq slice to create
the 3 masters in the install-config returns null for the missing
node.

Adding some validation here should make it clearer, but probably we should
catch this via validation in the installer.