08_deploy_bmo: Don't assume CLUSTER_PRO_IF == PRO_IF.
A new variable was added, CLUSTER_PROV_IF, which specifies which network
interface on the deployed hosts maps to the provisioning network.  The
script assumed that the value of PRO_IF is an appropriate default.

However, PRO_IF is the interface for the provisioning network on the
provisioning host.  We should not assume that network interfaces will
be the same between the provisioning host and the deployed hosts, so
separate these two options.