assets: Drop ingress-controller manifest.
The intention of this manifest was to change the node selector so that
the ingress controller would run on masters.  This manifest is owned
by the ingress operator, so our changes would just get overwritten.
In my local cluster, that does appear to be the case and there's no
node selector setting at all.