Remove facet references.
Facet development has been dormant for quite a while.  Some of the
relevant code was commented out, but we were still installing a few
unnecessary dependencies (yarn, nodejs).

Drop the facet related references from dev-scripts for now.  It can be
re-introduced when it's being actively used and we're sure the
additions are working properly.

When development of an installer UI resumes, it may take a different
form or be in a different location, so let's revisit dev-scripts
additions when that time comes.