Move waiting for worker to CI.
We should no longer set IPs on the worker Machine objects, as
cluster-api-provider-baremetal should be doing that automatically.

Since we're not setting IPs, there's no requirement for the script to
wait for the worker to come up, either.

This is still a useful verification, so move it to run_ci.sh, instead.

Related to issue #260