add BareMetalHost CRD
Until the machine-api-operator integration is finished we need the
host CRD to be defined during the installation process so the
installer can register host objects with the new cluster. This change
adds a copy of the CRD to the manifest templates so it is uploaded to
the cluster with the other manifests managed by the installer.