link machine to node by adding data to host
The actuator will copy the address and hostname information from the
host object onto the machine, which allows the machine to be linked to
the node. If we edit the machine directly, the actuator overwrites it
with empty data, so edit the host instead.

We now need to wait for the hosts to be registered in order to link
them to machines and nodes, so do this step last.