Change the order of the ironic containers
To ensure the IP address on the provisioning interface is set
We need static-ip-refresh to start before the others, observation
suggests the order in this file is the order they are started in.

Also move the BMO to the bottom of the list so that it can't
attempt to use ironic before its started.