Added a timeout on the wait for build in build_installer.sh
The default configuration (with no timeout) means that the script
to build a custom release can exit before completion due to the
default timeout being hit. This may not be an issue on a fast
environment where the build can come up quickly, but on a slightly
slower environment a timeout is wise to have.