Drop registry manifest.
The cluster-image-registry-operator will now use emptyDir by default
for the "baremetal" infrastructure platform, so this manifest is no
longer needed.

https://github.com/openshift/cluster-image-registry-operator/pull/332

Note this is still not a supported configuration, and is only used to
bring up the cluster.  It should be immediately reconfigured to use a
PVC once storage has been set up.  For more info on that post-install
task, see:

https://github.com/openshift-kni/install-scripts/issues/5

Closes #704