Don't rely on build date for `oc`
Recent 4.4 CI builds are missing the buildDate and it's causing errors
in dev-scripts. We're mostly interested that the initial `oc` client
is the specified version, the buildDate check was introduced during 4.3
to ensure our build had a specific fix.

Note: the recent CI builds are *also* missing gitVersion, but at least
this doesn't cause a problem when it's blank, other than forcing a new
client download.