Always use local release with MIRROR_IMAGES
Previously [0] would set the release override to the ci image instead
of the local one that we need to use when mirroring. This change moves
the logic for setting it into a common location where the image
override will be set correctly regardless of what combination of env
vars is in use.

0: https://github.com/openshift-metal3/dev-scripts/blob/1536fdeaeac964ecc27e876ec3e509e09563b93d/common.sh#L82