Run ifup on provisioning bridge after bouncing interface
On RHEL 8, when ifdown is run on a bridge's only (or last up) interface,
then the bridge is deleted. However, when ifup is run on the bridge's
interface, it is not correspondingly run on the bridge itself.

See:
https://github.com/fedora-sysv/initscripts/blob/rhel8-branch/network-scripts/ifdown-eth#L144

Since the provisioning interface is bounced with ifup/ifdown after
bringing the bridge up, then the bridge itself ends up not existing.
This patch adds an additional call to ifup the provisioning bridge after
bouncing the interface.

Signed-off-by: James Slagle <james.slagle@gmail.com>