Fix world-readable check in sanitychecks.sh
This issue was mentioned in #900 but I missed it when
merging - we need to ensure world-readable access to
the entire path, not only the $WORKING_DIR which may
be a subdirectory.

Also made the script executable and moved a variable to
enable easier local testing, and reversed the order of
the size and access test (the size test will always fail
if the directory is not readable, so we should check for
access first).