Add support for coredns-mdns to run_ci.sh
coredns-mdns is somewhat unique in that it doesn't become its own
image, it gets included in the coredns image as a vendored plugin.
Because of this, we need a bit more logic around building proposed
changes than just exporting the appropriate local image env var.

This change adds the ability to override the Dockerfile name for the
image build in dev-scripts (because coredns is also unusual in that
way) and adds the necessary steps to clone and update the coredns
code so it can be built with the proposed change.