support better per-cluster naming
This set of changes is the first in a series to support running
dev-scripts multiple times on the same host to create multiple
clusters. With these changes in place, the files created end up either
in cluster-specific locations or with the cluster name in the file
name. It is also possible to control the names of the provisioning and
baremetal networks using variables in the config file and control the
starting value for the VMBC ports used by the emulator.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>