Wire in hostname format parameters
Support was added in metal3-dev-env[0] for using custom hostname
formats for the Ironic nodes. This adds the necessary configuration
variables to dev-scripts to override the master and worker hostnames.
Real deployments frequently won't use master-X hostnames so it is
important that we can test those formats in a virtual environment.

0: https://github.com/metal3-io/metal3-dev-env/commit/3cf4203ffeb66a30bf5bf71efc462e7b8adbeb9f