Retrieve installer sha from CI environment
If we're in OpenShift CI in an installer PR, then the installer is
built by rebasing on the target branch (i.e. master). The SHA that
`openshift-install version` reports doesn't exist on GitHub. It's only
in the environment that CI used to build the installer.

This gets the PR SHA from the environment, and then fetchs the rhcos
from that branch. There's a potential problem here if a PR branch is out
of date. However, we really do need to get the rhcos.json from the
PR branch in case it's being updated.

The fixes for this are probably a bunch of API calls to see if the PR
modifies rhcos.json, or actually check out the git repo which will
increase the already long build time.