Ensure same release image is used for mirror and install
If a new release image is published between when the local mirror is
created and when the install-config file is created, it is possible
that the releases used may not match. This causes the local mirror
to not be used, which results in either unnecessary external traffic
or complete failure, depending on the situation.

This change writes the mirrored release image to a temp file so it
can be retrieved later when install-config is generated. This way
we know the same release image is used for both.

Closes #1001