Add dual stack IPv4+IPv6 support
This commit allows configuring a dual-stack IPv4+IPv6 environment.  It
changes a bunch of configuration to reflect that IPv4 and IPv6
settings are required at the same time, creates a dual-stack
"baremetal" network environment, and configures the OpenShift cluster
with both IPv4 and IPv6 subnets.

This isn't actually expected to result in a working OpenShift cluster
yet, but this provides a development and test environment for
dual-stack OpenShift clusters.