Add dual stack IPv4+IPv6 support (#1017)
* Fix libvirt network spec for latest metal3-dev-env

Update dev-scripts to specify the baremetal network subnets for IPv4
and IPv6 separate to be compatible with metal3-dev-env PR
https://github.com/metal3-io/metal3-dev-env/pull/280.

* Update metal3-dev-env pinned commit

Update metal3-dev-env to latest master commit, which includes the
dual-stack IPv6 support

* Add dual stack IPv4+IPv6 support

This commit allows configuring a dual-stack IPv4+IPv6 environment.  It
changes a bunch of configuration to reflect that IPv4 and IPv6
settings are required at the same time, creates a dual-stack
"baremetal" network environment, and configures the OpenShift cluster
with both IPv4 and IPv6 subnets.

This isn't actually expected to result in a working OpenShift cluster
yet, but this provides a development and test environment for
dual-stack OpenShift clusters.

* install-config: Switch to MachineNetwork

Switch from the deprecated MachineCIDR to MachineNetwork.  The new
field is a list and is required for dual-stack.