Only pin metal3-dev-env when not using a local checkout
When someone sets METAL3_DEV_ENV it most likely means they want
to test a local checkout so lets not reset in this case.

Also move the pin to 23eab39fe9977994aeac7ab6af603beb52059b32
which is the latest at the time of testing this PR and works
locally for me since #1017 merged.